+++ src/Amido.Stacks.Messaging.AWS.SNS.Tests/Amido.Stacks.Messaging.AWS.SNS.Tests.csproj
