+++ src/Amido.Stacks.Messaging.AWS.SNS.sln
