+++ src/Amido.Stacks.Messaging.AWS.SNS/Amido.Stacks.Messaging.AWS.SNS.csproj
