+++ src/Amido.Stacks.Messaging.AWS.SNS/AwsSnsConfiguration.cs
