+++ src/Amido.Stacks.Messaging.AWS.SNS/Events/Events.cs
