+++ src/Amido.Stacks.Messaging.AWS.SNS/Extensions/ServiceCollectionExtensions.cs
