+++ src/Amido.Stacks.Messaging.AWS.SNS/Logging/LogDefinition.cs
