+++ src/Amido.Stacks.Messaging.AWS.SNS/Publisher/EventPublisher.cs
